+++ package.json
@@ -5,7 +5,7 @@
   "dependencies": {
     "axios": "0.27.2",
     "classnames": "2.3.1",
-    "dayjs": "^1.11.5",
+    "dayjs": "^1.11.6",
     "firebase": "^9.9.2",
     "mobx": "6.6.1",
     "mobx-react-lite": "3.4.0",
