+++ src/common/constants/constants.ts
@@ -3,7 +3,6 @@ export const PAGES_PATHS = {
   SIGN_UP: '/signup',
   HOME: '/',
   RESET_PASS: '/reset-password',
-  PROFILE: '/profile',
 }
 
 export const BUTTON_LABELS = {
