+++ src/features/profile/components/profile-edit-name/profile-edit-name.module.scss
@@ -1,18 +1,18 @@
 @import 'styles/convert.scss';
 
-.editNameContent {
+.profileEditName {
   display: flex;
   flex-direction: column;
   gap: rem(20px);
 }
 
-.editNameForm {
+.profileEditNameForm {
   width: 100%;
   display: flex;
   gap: rem(20px);
 }
 
-.editNameInput {
+.profileEditNameInput {
   width: 100%;
   display: flex;
   flex-direction: column;
