+++ src/features/profile/components/profile-label/profile-label.tsx
@@ -1,9 +1,15 @@
+import { FC, ReactNode } from 'react'
+
 import styles from './profile-label.module.scss'
 
-export const ProfileLabel = ({ children }: any) => {
+interface IProps {
+  children: ReactNode
+}
+
+export const ProfileLabel: FC<IProps> = ({ children }) => {
   return (
-    <div className={styles.profileContent}>
-      <div className={styles.profileLabel}>{children}</div>
+    <div className={styles.profileLabel}>
+      <div className={styles.profileLabelChildren}>{children}</div>
     </div>
   )
 }
