+++ src/pages/page-router.tsx
@@ -5,8 +5,6 @@ import { PageAuth } from 'pages/page-auth/page-auth'
 
 import { PAGES_PATHS } from 'common/constants/constants'
 
-import { PageUser } from './page-user/page-user'
-
 export const PageRouter = () => {
   return (
     <BrowserRouter>
@@ -15,7 +13,6 @@ export const PageRouter = () => {
         <Route path={PAGES_PATHS.LOG_IN} element={<PageAuth type='login' />} />
         <Route path={PAGES_PATHS.SIGN_UP} element={<PageAuth type='signup' />} />
         <Route path={PAGES_PATHS.RESET_PASS} element={<PageAuth type='reset' />} />
-        <Route path={PAGES_PATHS.PROFILE} element={<PageUser />} />
       </Routes>
     </BrowserRouter>
   )
