--- src/pages/page-user/page-user.module.scss
@@ -1,21 +1,0 @@
-@import 'styles/convert.scss';
-
-.pageUser {
-  margin: 0 auto;
-  padding: rem(32px) 0 0 0;
-  display: flex;
-  max-width: rem(1260px);
-  gap: rem(128px);
-  justify-content: center;
-}
-
-.pageUserStepper {
-  display: flex;
-  flex-direction: column;
-  gap: rem(30px);
-  width: rem(165px);
-}
-
-.pageUserContent {
-  width: rem(652px);
-}
